fix: sync with light engine before saving
because the chunk system gets too fast

When chunks with light maps get deserialized, the process of
loading light data is asynchronous. But the chunk is assumed to have
lighting on it. The process of loading lighting data isn't atomic,
so there is a good chance to get incomplete light maps from the
light engine while it is loading the lighting data.

In the context of chunk system, if the chunk never reaches
`minecraft:initialize_light`, where synchronization with the
light engine happens, and proceeds to be completely unloaded,
light maps can be broken:
- If saving happens before the light data starts loading in,
  the light map will be completely gone.
- If saving happens during the process of loading the light data,
  the light map will be incomplete.

This commit adds an additional synchronization with the light engine
to mitigate such issue.